misc/HtmlTestSuite/results/test-1143.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
    <title>'self' keyword positive test</title>
    <meta content="script-src 'self' 'nonce-abc';" http-equiv="Content-Security-Policy"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
    <script nonce="abc">
      var t_spv = async_test("Should fire violation events for every failed violation");
      window.addEventListener(
        "securitypolicyviolation", t_spv.unreached_func("securitypolicyviolation should not be emitted"));

      window.addEventListener("load", function() {
        t_spv.done();
      });
    </script>
    <script src="positiveTest.js"/>
    <script nonce="abc">
      test(function() {
        assert_true(window.cspPositiveTest);
      }, "Allows scripts from the same host.");
    </script>
</head>
<body>
    <h1>'self' keyword positive test</h1>
    <div id="log"/>
</body>
</html>
